organizando melhor o código
